services/llm/requirements.txt
flask
werkzeug
pillow
numpy
langchain
langchain-openai
pydantic